+++ tests/unit/util/viewpoint.spec.ts
@@ -1,3 +1,4 @@
+import { expect } from 'chai';
 import Viewpoint from '../../../src/util/viewpoint.js';
 
 describe('Viewpoint', () => {
@@ -13,6 +14,31 @@ describe('Viewpoint', () => {
         animate: false,
       });
       expect(vp.isValid()).to.be.false;
+    });
+
+    it('should be invalid if coordinates are out of bounds', () => {
+      expect(
+        new Viewpoint({
+          name: 'hauptbhnhf',
+          distance: 2480.9491308407964,
+          groundPosition: [190, 30, 3],
+          heading: 337.52854532052663,
+          pitch: -41.10601490149314,
+          roll: 359.8987925198945,
+          animate: false,
+        }).isValid(),
+      ).to.be.false;
+      expect(
+        new Viewpoint({
+          name: 'hauptbhnhf',
+          distance: 2480.9491308407964,
+          cameraPosition: [80, 100, 3],
+          heading: 337.52854532052663,
+          pitch: -41.10601490149314,
+          roll: 359.8987925198945,
+          animate: false,
+        }).isValid(),
+      ).to.be.false;
     });
 
     it('should be invalid on non-numeric axis parameters', () => {
@@ -31,6 +57,8 @@ describe('Viewpoint', () => {
         animate: false,
       });
       expect(vp.isValid()).to.be.true;
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       vp.roll = '123';
       expect(vp.isValid()).to.be.false;
     });
@@ -53,8 +81,8 @@ describe('Viewpoint', () => {
   });
 
   describe('comparing viewpoints', () => {
-    let vp1;
-    let vp2;
+    let vp1: Viewpoint;
+    let vp2: Viewpoint;
 
     beforeEach(() => {
       vp1 = new Viewpoint({
@@ -97,12 +125,12 @@ describe('Viewpoint', () => {
     });
 
     it('should return true, if the viewpoints values are equal within epsilon', () => {
-      vp2.distance += 0.01;
+      vp2.distance! += 0.01;
       vp2.heading += 0.01;
       vp2.pitch -= 0.01;
       vp2.roll -= 0.01;
-      vp2.groundPosition[0] += 0.01;
-      vp2.cameraPosition[0] += 0.01;
+      vp2.groundPosition![0] += 0.01;
+      vp2.cameraPosition![0] += 0.01;
       expect(vp1.equals(vp2, 0.1)).to.be.true;
       expect(vp1.equals(vp2, 0.05)).to.be.true;
     });
@@ -117,26 +145,28 @@ describe('Viewpoint', () => {
     });
 
     it('should return false, if other viewpoint is null', () => {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       expect(vp1.equals(null)).to.be.false;
     });
 
     it('should return false, if the viewpoints values are NOT equal', () => {
-      vp2.distance += 0.01;
+      vp2.distance! += 0.01;
       vp2.heading += 0.01;
       vp2.pitch -= 0.01;
       vp2.roll -= 0.01;
-      vp2.groundPosition[0] += 0.01;
-      vp2.cameraPosition[0] += 0.01;
+      vp2.groundPosition![0] += 0.01;
+      vp2.cameraPosition![0] += 0.01;
       expect(vp1.equals(vp2)).to.be.false;
     });
 
     it('should return false, if the viewpoints values are NOT equal within epsilon', () => {
-      vp2.distance += 0.01;
+      vp2.distance! += 0.01;
       vp2.heading += 0.01;
       vp2.pitch -= 0.01;
       vp2.roll -= 0.01;
-      vp2.groundPosition[0] += 0.01;
-      vp2.cameraPosition[0] += 0.01;
+      vp2.groundPosition![0] += 0.01;
+      vp2.cameraPosition![0] += 0.01;
       expect(vp1.equals(vp2, 0.01)).to.be.false;
       expect(vp1.equals(vp2, 0.001)).to.be.false;
     });